Core Community Coop Files
oAuth has been edited out as it is a private stream key. This is the foundation of an api chat based game that engages the community to play and interact with on my desktop application. Some tester/non-functional scripts and assets but will remove when officially published. Includes a tester text file for data collection of player behaviour/commands